+++ kifi-backend/modules/curator/app/com/keepit/curator/commanders/RecommendationGenerationCommander.scala
@@ -151,39 +151,36 @@ class RecommendationGenerationCommander @Inject() (
               }
             }
 
-            val scoredItemsFut = scoringHelper(cleanedItems)
-            val toBeSavedItemsFut = scoredItemsFut.map { scoredItems => scoredItems.filter(si => shouldInclude(si.uriScores)) }
-            val attributionFut = toBeSavedItemsFut.flatMap { items => attributionHelper.getSeedAttribution(items) }
+            val toBeSaved = scoringHelper(cleanedItems).map { scoredItems =>
+              scoredItems.filter(si => shouldInclude(si.uriScores))
+            }.flatMap { scoredItems =>
+              attributionHelper.getAttributions(scoredItems)
+            }
 
-            for {
-              toBeSavedItems <- toBeSavedItemsFut
-              attribution <- attributionFut
-            } yield {
+            toBeSaved.map { items =>
               db.readWrite { implicit s =>
-                (toBeSavedItems zip attribution) foreach {
-                  case (item, attr) =>
-                    toBeSavedItems.map { item =>
-                      val recoOpt = uriRecRepo.getByUriAndUserId(item.uriId, userId, None)
-                      recoOpt.map { reco =>
-                        uriRecRepo.save(reco.copy(
-                          masterScore = computeMasterScore(item.uriScores),
-                          allScores = item.uriScores,
-                          attribution = attr
-                        ))
-                      } getOrElse {
-                        uriRecRepo.save(UriRecommendation(
-                          uriId = item.uriId,
-                          userId = userId,
-                          masterScore = computeMasterScore(item.uriScores),
-                          allScores = item.uriScores,
-                          seen = false,
-                          clicked = false,
-                          kept = false,
-                          attribution = attr
-                        ))
-                      }
-                    }
+                items foreach { item =>
+                  val recoOpt = uriRecRepo.getByUriAndUserId(item.uriId, userId, None)
+                  recoOpt.map { reco =>
+                    uriRecRepo.save(reco.copy(
+                      masterScore = computeMasterScore(item.uriScores),
+                      allScores = item.uriScores,
+                      attribution = item.attribution
+                    ))
+                  } getOrElse {
+                    uriRecRepo.save(UriRecommendation(
+                      uriId = item.uriId,
+                      userId = userId,
+                      masterScore = computeMasterScore(item.uriScores),
+                      allScores = item.uriScores,
+                      seen = false,
+                      clicked = false,
+                      kept = false,
+                      attribution = item.attribution
+                    ))
+                  }
                 }
+
                 genStateRepo.save(newState)
               }
