+++ kifi-backend/modules/curator/app/com/keepit/curator/model/SeedItem.scala
@@ -1,7 +1,7 @@
 package com.keepit.curator.model
 
 import com.keepit.common.db.{ SequenceNumber, Id }
-import com.keepit.model.{ NormalizedURI, User }
+import com.keepit.model.{ Keep, NormalizedURI, User }
 import com.kifi.macros.json
 
 import org.joda.time.DateTime
@@ -36,3 +36,14 @@ case class SeedItem(
 }
 
 case class ScoredSeedItem(userId: Id[User], uriId: Id[NormalizedURI], uriScores: UriScores)
+
+@json case class UserAttribution(friends: Seq[Id[User]], others: Int)
+@json case class KeepAttribution(keeps: Seq[Id[Keep]])
+@json case class TopicAttribution(topicName: String)
+@json case class SeedAttribution(user: Option[UserAttribution] = None, keep: Option[KeepAttribution] = None, topic: Option[TopicAttribution] = None)
+
+object SeedAttribution {
+  val EMPTY = SeedAttribution()
+}
+
+case class ScoredSeedItemWithAttribution(userId: Id[User], uriId: Id[NormalizedURI], uriScores: UriScores, attribution: SeedAttribution)
